decouple proof related stuff
